Removed mistakenly identified section.
Removed section of event packing responsible for bitRegister and LAM
masks as this turned out to be a Caen V977 IO Register. Currently the
code as a hard coded section to deal with this module that should be
removed and rewritten. The new paradigm will require the end user to
specifically which modules they have and which order to unpack them. Each
module will need a class to handle unpacking.