Final revisions for evt2root3.
Updated evt2root to print buffers less frequently if using Ring Buffer.
Separated the event length routine from DumpEvent and ReadEvent. Added
some protection for reading events that are longer than the buffer
containing them. If this is true the buffer is fast forwarded to the end
since we must be lost. evt2root and evtDump not check that the buffer is
not at the end before requesting another read.