Initial commit of version 3.
This commit is not stable and should not be used for production.

Separted the nscl buffer class into class for each type of system. Created an abstract base class which all buffer classes are derived. Added a new buffer class for files from HHIRF L003 formats. Created an abstraact module class from which all modules are derived. Changes in the way the eventData and eventScaler classes are called. Updated cmake module and build.