Generalizations, and extension of nsclRingBuffer.
The evtDump was generalized so that a recompile is not needed when switching between buffer classes. This should be continued such that a configuration file can be specified containing the modules from the data buffer.

nsclRingBuffer extended further to v11. mainBuffer accounting was changed to bytes to deal with noninteger word sized buffers.

All buffer classes are now required to have an UnpackBuffer() method to do tasks specific to that buffer type. Currently for evt2root to work both the physics data buffers and scaler buffers are ignored by this method and dealt with manually.