Branches can now be created from modules.
We now generate a ROOT dictionary of all the modules such that tree branches of a module can be created. This will allow a default raw tree to be built if a sorting library is not specified. This required that baseModule inherited from TObject and all modules have ClassDef and ClassImp as well as the addition of a LinkDef file for the modules.