Corrected bug in hribfModule.
The hribfModule ran a loop searching for a -1 trailer word, which was not always created. We now loop on the number of bytes left in a buffer. The module takes care of the trailer and this message was removed from hribfBuffer::ReadEvent. The peeking to inspect if there are anymore events occurs by first checking htat there are bytes remaining. This loop was infinite as the mainBuffer::GetWord() method was permitting a word to be read as the buffer size had not been reached, but the numOfBytes to be read was set to 0, thus returning a 0 word. GetWord() now checks against the number of valid bytes instead of the buffer size.