evtDump updated to dump events and additional flag.
evtDump was not show event dumps as the verobisty was not set to true.

Added additional option to skip a fixed number of buffers in a file. Useful when an error is generated part way through a file.

Added setup for configuring modules as a hardcoded line this should be replaced by a config file.