Added IsRunEnd and some bug fixes.
Added IsRunEnd() method to mainBuffer and all descendants to allow checking in evt2root. evt2root should be moved to use UnpackBuffer() method.

Corrected GetNumberEventsRemaining to return a positive value.

Added an unknown option message prior to usage call in evt2root. Allowed the use to not specify a module list which will be used by a future format.