Corrected multijob compilation.
The root_generate_dictionary was being called multiple times
simoltaneously to build the static and shared module libraries. This has
been resolved by now building objects and then using those unlinked
objects to build the libraries. The building of objects generates the
root library only once.

This required cmake 2.8.8 and requires that the objects are compiled as
Position Independent Code.