Corrected fast list data unpacking order.
FAST list buffers were not being unpacked int he proper order. The dummy
word had caused confusion about the order. The dummy word is placed at
the beginning of the list of data not at the end. Assuming it was at the
end led to a misunderstanding of the data's endianness thus swap the ADC
values between data neighboring ADCs.

A check of the 31st bit if the number of ADCs was odd was added as well
as a reduction in the compexity of the ReadEvent.