Added multiplicity information for HRIBF format.
The HRIBF format may have data events where a parameter (channel) is set
multiple times. The meaning of this may be ambiguous, but perhaps is
expected, e.g. multi-hit TDCs. This commit adds a variable mult which
tracks the number of parameters read out and a vector paramMults which
tracks the number of times each parameter is read out. Currently the
last parameter data readout will be stored. Due to this an error message
is printed for each event with this situation listing the buffer, event
number and parameters that were multiply hit.