Added SRC_ID and additional checks GetWord checks.
The main buffer now will store the source ID associated with each
module. This is specifically designed for the nsclRing buffer, but could
be a general feature for any type of buffer. The ring buffer allows
multiple acq to be stored into one format and each one may have a
different set of modules. The buffer specifies the source ID during
unpacking and the module list to be unpacked is identified against this
number.

In addition, a couple of check in GetWord and GetCurrentWord were added
to reduce the number of errors that are produced when there fractional
number of words in a buffer and when a fractional word is read.