Added buffer source and multi file support.
The buffer source ID is now carried through evt2root and supported by
nsclRingBuffer. Revised the filling of the sourceID vector to fill with
-1 when a sourceID is not specified for a single module. If none is
specified at all then the default value is 0.

Added multi file support to evtDump so that two files can be scanned
together. This helps in causes such as the ring buffer where a buffer
changes the readout methods.